+++ frontend/src/components/InstagramPostComment.vue
@@ -1,29 +1,26 @@
 <template>
-  <div
-    class="instagram-post-comment rounded-lg"
-    style="direction: rtl"
+  <card
+    :variant="variant"
   >
-    <div class="d-flex flex-row justify-content-between w-100">
-      <p class="m-0"> <small> {{ username }} </small> </p>
-      <b-icon
-        :icon="sentiment === 'pos' ? 'emoji-smile' : sentiment === 'neg' ? 'emoji-frown' : sentiment === 'neu' ? 'emoji-neutral' : 'question-circle'"
-        :variant="sentiment === 'pos' ? 'success' : sentiment === 'neg' ? 'danger' : sentiment === 'neu' ? 'secondary' : 'dark'"
-        font-scale="3"
-      ></b-icon>
-    </div>
     <div
-      class="d-flex flex-row justify-content-start jumbotron border p-0 my-2"
-      :class="[sentiment === 'pos' ? 'border-success' : sentiment === 'neg' ? 'border-danger' : 'border-secondary']"
+      class="d-flex flex-row align-items-center w-100"
+      style="color: #5F5F5F;"
     >
-      <p class="p-3 my-2"> {{ text }} </p>
+      <p class="m-0 border-left"> <b-badge dir="ltr" variant="transparent"> {{ username }} </b-badge> </p>
+      <p class="m-0" dir="ltr"> {{ jalaliDate }} </p>
     </div>
-    <p class="m-0"> <small> {{ createdAt }} </small> </p>
-  </div>
+    <p class="p-3 my-2"> {{ text }} </p>
+  </card>
 </template>
 
 <script>
+import Card from '@/components/Card'
+
 export default {
   name: 'InstagramPostComment',
+  components: {
+    Card
+  },
   props: {
     createdAt: {
       type: Number,
@@ -47,6 +44,14 @@ export default {
       type: String,
       required: false,
       default: ''
+    }
+  },
+  computed: {
+    variant: function () {
+      return this.sentiment === 'pos' ? 'success' : this.sentiment === 'neg' ? 'danger' : 'info'
+    },
+    jalaliDate: function () {
+      return new Date(this.createdAt * 1000).toLocaleString('fa-IR')
     }
   }
 }