+++ frontend/src/mixin/index.js
@@ -2,6 +2,7 @@ import GeneralMixin from './general'
 import QueryMixin from './query'
 import FilterMixin from './filter'
 import TelegramMixin from './telegram'
+import InstagramMixin from './instagram'
 import TwitterMixin from './twitter'
 import KhabarFooriMixin from './web'
 
@@ -10,6 +11,7 @@ export {
   QueryMixin,
   FilterMixin,
   TelegramMixin,
+  InstagramMixin,
   TwitterMixin,
   KhabarFooriMixin
 }