+++ CHANGES.md
@@ -6,7 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-[1.0.0] - 2021-07-13
+## [1.2.0] - 2021-07-25
+### Fixed
+- Instagram Crawler (#24)
+- Telegram Channel username/title (#36)
+- Instagram jalali date (#39)
+
+## [1.0.0] - 2021-07-13
 ### Added
 - Changelog file
 - Docker compose